kanban-todo-api/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
psycopg2-binary==2.9.9

# JWT and security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0

# Testing tools
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
